refactor: separate navbar components into separate files
